Rename sprites.py --> model.py
